+++ nym-vpn-core/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - While in Connected state swap internal resolver to use custom DNS (via system resolver). (https://github.com/nymtech/nym-vpn-client/pull/5674)
 
+### Fixed
+
+- Improve forwarding resolver behavior by not sending empty response when hostname resolution fails. (https://github.com/nymtech/nym-vpn-client/pull/5832)
+
 
 ## [2026.11.0] - TBD